consolidate inits
